Fix new buffer slow loading for ruby in nvim
The previous fix made things worse with filetypes not being recognised
and comment strings and other local options not being set for buffers.

So we dont need to switch off the file type plugin instead the issue was
due to not proper setup of path vars for rvm. So the ruby path was
taking time.